agent/requirements.txt
# 毕业设计：LangChain / LangGraph 工作流
langgraph>=0.2.55
langchain>=0.3.14
langchain-openai>=0.2.14
langchain-core>=0.3.28
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
python-dotenv>=1.0.1
httpx>=0.28.0
pydantic>=2.10.0
